+++ packages/plugin-koa/test/koa.test.ts
@@ -34,7 +34,7 @@ describe('plugin: koa', () => {
   })
 
   describe('requestHandler', () => {
-    it('should start a session and attach a client to the context', async () => {
+    it('should clone the client, start a session and attach the cloned client to the context', async () => {
       const client = new Client({ apiKey: 'api_key', plugins: [plugin] })
 
       const startSession = jest.fn().mockReturnValue(client)
@@ -43,6 +43,7 @@ describe('plugin: koa', () => {
 
       client._sessionDelegate = { startSession, pauseSession, resumeSession }
       client._logger = logger()
+      client._clientContext = { run: jest.fn() }
 
       const middleware = client.getPlugin('koa')
 
@@ -57,11 +58,12 @@ describe('plugin: koa', () => {
 
       expect(client._logger.warn).not.toHaveBeenCalled()
       expect(client._logger.error).not.toHaveBeenCalled()
-      expect(startSession).not.toHaveBeenCalled()
+      expect(startSession).toHaveBeenCalledTimes(1)
       expect(pauseSession).not.toHaveBeenCalled()
-      expect(resumeSession).toHaveBeenCalledTimes(1)
-      expect(context.bugsnag).toBe(client)
-      expect(next).toHaveBeenCalledTimes(1)
+      expect(resumeSession).not.toHaveBeenCalled()
+      expect(context.bugsnag).toStrictEqual(expect.any(Client))
+      expect(context.bugsnag).not.toBe(client)
+      expect(client._clientContext.run).toHaveBeenCalledWith(expect.any(Client), next)
     })
 
     it('should record metadata from the request', async () => {
@@ -104,15 +106,16 @@ describe('plugin: koa', () => {
         },
         ip: '1.2.3.4'
       } as any
+      client._clientContext = { run: jest.fn() }
 
       const next = jest.fn()
 
       await middleware.requestHandler(context, next)
 
-      expect(next).toHaveBeenCalledTimes(1)
+      expect(client._clientContext.run).toHaveBeenCalledWith(expect.any(Client), next)
 
       const event: Event = await new Promise(resolve => {
-        client.notify(new Error('abc'), noop, (_, event) => resolve(event as Event))
+        (context.bugsnag as Client).notify(new Error('abc'), noop, (_, event) => resolve(event as Event))
       })
 
       expect(client._logger.warn).not.toHaveBeenCalled()
@@ -160,6 +163,7 @@ describe('plugin: koa', () => {
         },
         sendSession: noop
       }))
+      client._clientContext = { run: jest.fn() }
 
       const middleware = client.getPlugin('koa')
 
@@ -177,10 +181,10 @@ describe('plugin: koa', () => {
 
       await middleware.requestHandler(context, next)
 
-      expect(next).toHaveBeenCalledTimes(1)
+      expect(client._clientContext.run).toHaveBeenCalledWith(expect.any(Client), next)
 
       const event: Event = await new Promise(resolve => {
-        client.notify(new Error('abc'), noop, (_, event) => resolve(event as Event))
+        (context.bugsnag as Client).notify(new Error('abc'), noop, (_, event) => resolve(event as Event))
       })
 
       expect(client._logger.warn).not.toHaveBeenCalled()
@@ -218,6 +222,7 @@ describe('plugin: koa', () => {
 
       client._sessionDelegate = { startSession, pauseSession, resumeSession }
       client._logger = logger()
+      client._clientContext = { run: jest.fn() }
 
       const middleware = client.getPlugin('koa')
 
@@ -235,7 +240,7 @@ describe('plugin: koa', () => {
       expect(startSession).not.toHaveBeenCalled()
       expect(pauseSession).not.toHaveBeenCalled()
       expect(resumeSession).not.toHaveBeenCalled()
-      expect(next).toHaveBeenCalledTimes(1)
+      expect(client._clientContext.run).toHaveBeenCalledWith(expect.any(Client), next)
 
       // the Client should be cloned to ensure any manually started sessions
       // do not leak between requests