+++ packages/plugin-node-unhandled-rejection/unhandled-rejection.js
@@ -3,16 +3,20 @@ module.exports = {
   load: client => {
     if (!client._config.autoDetectErrors || !client._config.enabledErrorTypes.unhandledRejections) return
     _handler = err => {
-      const event = client.Event.create(err, false, {
+      // if we are in an async context, use the client from that context
+      const ctx = client._clientContext && client._clientContext.getStore()
+      const c = ctx || client
+
+      const event = c.Event.create(err, false, {
         severity: 'error',
         unhandled: true,
         severityReason: { type: 'unhandledPromiseRejection' }
       }, 'unhandledRejection handler', 1)
 
       return new Promise(resolve => {
-        client._notify(event, () => {}, (e, event) => {
-          if (e) client._logger.error('Failed to send event to Bugsnag')
-          client._config.onUnhandledRejection(err, event, client._logger)
+        c._notify(event, () => {}, (e, event) => {
+          if (e) c._logger.error('Failed to send event to Bugsnag')
+          c._config.onUnhandledRejection(err, event, c._logger)
           resolve()
         })
       })